# service/src/main/java/com/spbstu/edu/advertisement/dto/FavouriteDto.java
package com.spbstu.edu.advertisement.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Builder
@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
public class FavouriteDto {
    private Long adId;

    private Boolean isFavourite;
}